% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{avermiss}
\alias{avermiss}
\title{Performance of lasso}
\usage{
avermiss(X, beta, sigma, N)
}
\arguments{
\item{X}{the independent variables in the linear regression model}

\item{beta}{the coefficient}

\item{sigma}{the standard error of epsilon}

\item{N}{the number of simulation}
}
\value{
the average number of regressors missed from the true model,the average number of regressors selected outside the model,relative average empirical risk
}
\description{
A function to measure the performance of lasso in high-dimensioanl linear regression model
}
\examples{
\dontrun{
library(MASS)
n=100
p=500
sigma <- matrix(numeric(p^2),nrow=p)
for(i in 1:p){
  for(j in 1:p){
    if(i==j){sigma[i,j]<-1}
    else{sigma[i,j]<- (1/2)^abs(i-j)}
  }
}
mean <- numeric(p)
set.seed(1234)
x <- mvrnorm(n,mean,sigma)
beta <- numeric(p)
beta[1:5]=1
avermiss(x,beta,1,100)
 }
}
