% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\arguments{
\item{n}{the number of sample}

\item{n1}{the number of between-sample random numbers}

\item{a}{a constant}

\item{b}{a constant}
}
\value{
a random sample of size n
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
n <- 1e4 #length of chain
a <- 2
b <- 2
n1=10
mtrix <- matrix(0,n,2)
mtrix <- gibbsC(n,n1,a,b)
} 
}
