% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrootlasso.R
\name{sqrootlasso}
\alias{sqrootlasso}
\title{Square-root lasso}
\usage{
sqrootlasso(x, y, lambda, maxIter)
}
\arguments{
\item{x}{the independent variables in the linear regression model}

\item{y}{the response variable in the linear regression model}

\item{lambda}{the penalty factor}

\item{maxIter}{the maximum number of iterations}
}
\value{
beta a estimate of regression coefficient
}
\description{
A function to estimate high-dimensional sparse linear regression model
}
\examples{
\dontrun{
library(MASS)
n=100
p=500
sigma <- matrix(numeric(p^2),nrow=p)
 for(i in 1:p){
 for(j in 1:p){
 if(i==j){sigma[i,j]<-1}
 else{sigma[i,j]<- (1/2)^abs(i-j)}}}
 mean <- numeric(p)
 set.seed(1234)
  x <- mvrnorm(n,mean,sigma)
  beta <- numeric(p)
  beta[1:5]=1
  epsilon=rnorm(n) 
  y=x\%*\%beta+epsilon  
  b <- sqrootlasso(x,y,42.7,1e4)
  }
}
