\name{StatComp21052-package}
\alias{StatComp21052-package}
\alias{StatComp21052}
\docType{package}
\title{
  This package provides two functions to solve square-root lasso,and      measure the performance of lasso.
}
\description{
  The "sqrootlasso" function is to estimate the regression coefficient    by using square-root lasso in high-dimensional linear model.The         "avermiss" function is to measure the performance off lasso by          calculating the average number of regressors missed from the true       model ,the average number of regressors selected outside the model and   relative average empirical risk.
}
\author{
Xiaoyan Chen

Maintainer: Xiaoyan Chen <1641207431@qq.com>
}
\keyword{ package }
